# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FutureGroup",
        "contact": {
            "email": "sohibjon.developer@mail.ru"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/notebook": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Notes",
                "operationId": "getNotebooksList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "Notebook::class"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Post note",
                "description": "Post note",
                "operationId": "PostNotebook",
                "parameters": [
                    {
                        "name": "fio",
                        "in": "query",
                        "description": "Fio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company",
                        "in": "query",
                        "description": "Company",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "phone",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "born_date",
                        "in": "query",
                        "description": "born date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo"
                                ],
                                "properties": {
                                    "photo": {
                                        "description": "Photo",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "Notebook::class"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notebook/{id}": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "note by ID",
                "description": "single note",
                "operationId": "getNotebookById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of note",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "Notebook::class"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Update note",
                "description": "Update ",
                "operationId": "updateNote",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "fio": {
                                        "description": "New FIO",
                                        "type": "string"
                                    },
                                    "company": {
                                        "description": "New company",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "New phone",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "New Email",
                                        "type": "string"
                                    },
                                    "born_date": {
                                        "description": "New Born date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "photo": {
                                        "description": "New photo",
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notebook not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notebook"
                ],
                "summary": "delete note by ID",
                "description": "delete note",
                "operationId": "deleteNote",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of note",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "Notebook::class"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            }
        }
    }
}